refactor query embedding generation
